Remove sourcever from build files.
When the code was Java 1.4, it helped to silence a few
warnings. Now the code is Java 5, but with OpenJDK 9 source
version 5 is no longer supported and version 6 gives a warning
about being deprecated. Seems that it's simplest to forget
the source version option and just be careful to avoid
introducing new language features unintentionally.