Restore the -source option and add also an explicit -target.
Removing the source version was a big mistake because setting
the source version also implicitly set the target version.
I had forgotten this. When the -source option was removed,
the resulting binaries ran only on a recent JDK version.

This commit restores the -source option and also adds an
explicit -target that is always set to the same value as
the -source. In this commit it's set to Java 5 even though
OpenJDK 8 is the last one that supports it.

Thanks to Brett Okken and Stefan Bodewig.