Speed optimization to LZDecoder.repeat.
With extremely compressible data this may even double the LZMA2
decompression speed. However, with typical files the improvement
is minor (possibly only 1-2 %).

Big thanks to Brett Okken. He suggested using a loop where the
amount to copy doubles on each iteration which is a great idea
in a Java implementation of this method. He also took care of
benchmarking several different alternatives.