Fixing building of the docs.
The linkoffline attribute in Ant 1.10.9 only looks for the file
"package-list" while Java >= 11 use "element-list" instead.
To workaround this problem in Ant, rename element-list to
package-list. This seems to work fine with Java 15 but obviously
it's a hack that might break some day.

Don't specify the source language version when building the docs.
If source="7" is used and linking to Java 15 documentation that
uses element-list, this warning will be repeated for every file:

    The code being documented uses packages in the
    unnamed module, but the packages defined in
    https://docs.oracle.com/en/java/javase/15/docs/api/
    are in named modules.