improve byte array mismatch calculation performance
Much of the compression work is comparing ranges of binary values. These
changes improve the performance for every platform on jdk9+.

For x86 this is done by using a byteArrayViewVarHandle to process 8 (64
bit) or 4 bytes at a time.

https://docs.oracle.com/javase%2F9%2Fdocs%2Fapi%2F%2F/java/lang/invoke/MethodHandles.html#byteArrayViewVarHandle-java.lang.Class-java.nio.ByteOrder-

For all other platforms, this uses the Arrays.mismatch method to do the
comparison, which is implemented in the jdk itself with vectors.

https://docs.oracle.com/javase%2F9%2Fdocs%2Fapi%2F%2F/java/util/Arrays.html#mismatch-byte:A-int-int-byte:A-int-int-

For older jdks (7 and 8) on x86 or aarch64, sun.misc.Unsafe is used (if
it can be found) to process 8 (64 bit) or 4 bytes at a time.

As previously discussed on the mailing list, this shows substantial
improvements (~20%) in total compression times for a variety of
real-world data sets for both arm and x86, with only minor regressions
if forced (by system property) to legacy byte-by-byte comparisons.
https://www.mail-archive.com/xz-devel@tukaani.org/msg00400.html