Switch to BSD Zero Clause License.
This adds AUTHORS as COPYING now contains the copyright notice
and the license text. SPDX identifiers were added to most files.

The switch to 0BSD is done because public domain has
(real or perceived) legal issues in some jurisdictions.
All people mentioned in AUTHORS have confirmed that
their works included in this project can be distributed
under the 0BSD license.