Delta filter: Improve performance.
Now an unneeded roundtrip via the history[] buffer is avoided
for most bytes. Assuming that the input and output buffers
are at least a few kilobytes, this can reduce Delta encoding time
by 60-70 % when distance=1 and over 50 % with long distances.

Decoding time reduces up to 50 % with long distances but only
5 % with distance=1. In any case it's always an improvement.

With the old code, different distance values didn't affect
encoding or decoding speed.

However, in .xz files together with LZMA2, the LZMA2 speed
dominates and the benefit of the Delta filter improvement is
very tiny. It's still nice to have as the code is still simple.

Closes: https://github.com/tukaani-project/xz-java/pull/11

(This commit message was written by Lasse Collin.)