Add VarHandle-based CRC64 for Java >= 9
Now the slicing-by-4 is done properly in terms of performance
since it reads 32-bit integers at a time.

This is 20-25 % faster on x86-64. It should be faster on all
archs and it's compatible with strict-align archs too. In
contrast to XZ Utils, this implementation is little-endian only
to keep it simpler.

Good speed on big endian requires that byteswap in VarHandle is
fast. Testing this code on x86-64 with getIntBE, the difference
is maybe 1 % or maybe no difference at all. So on x86-64 it
doesn't matter much which endianness is used. Hopefully it's
like that on the few big endian systems with Java too.

A separate big endian version might be considered *if* it seems valuable
enough and that someone can also properly test it on real hardware.

The flattening of TABLE[] helped here. Testing it on the pre-VarHandle
code, it didn't matter much with OpenJDK 21 (maybe a tiny bit faster)
but it was slower on Java 8. Now that the old code is for Java 8 only,
it seems best to leave it unchanged.