Update README and convert to README.md
I find it important that the plain text look and readability stays
as good as possible because the files are read as plain text files
without a CommonMark compatible Markdown renderer too. It's only
slightly more effort to edit the files this way.

Don't change the README reference in NEWS.md because 1.9 has README,
not README.md.

Suggested-by: GitHub user JNightRider
Closes: https://github.com/tukaani-project/xz-java/pull/17