Fix this-escape warning in XZOutputStream
Don't call the public method updateFilters(FilterOptions[]) from
a constructor. If a subclass had overridden this method,
XZOutputStream's constructor would call the method in the subclass,
which would lead to "this" escaping before the subclass has been
initialized.

This commit likely counts as an API/ABI breaking change but in practice
it matters only if someone has extended the class and relied on this
broken behavior. Existence of such a subclass sounds highly unlikely
in practice though.