XZOutputStream: Don't call updateFilters from constructors
Fix a this-escape warning in XZOutputStream by not calling the public
method updateFilters(FilterOptions[]) from constructors. If a subclass
had overridden the method, XZOutputStream's constructors would call
the method in the subclass, which would lead to "this" escaping before
the subclass had been initialized. The API docs didn't mention that
such a call was being made by the constructors.

This is an API/ABI breaking change but in practice it matters only if
someone has extended the class and overridden this rarely-used method
in a way that relies on the undocumented behavior of the constructors.
Existence of such a subclass sounds highly unlikely in practice.

Thanks to Brett Okken for the discussion.
See: https://github.com/tukaani-project/xz-java/pull/18