Make all filter options classes final
This fixes this-escape warnings in DeltaOptions and LZMA2Options.
The other classes were changed for consistency.

This is an API/ABI breaking change which matters if someone has
extended these classes. I didn't find code on the web that had done
so, thus it feels likely that few (if any) users of XZ for Java will
be affected. This commit can be reconsidered if needed though.

The filter options classes validate the options. Other classes assume
that the getter methods return valid values. So a very minor extra
benefit is that now there cannot be subclasses that could break the
assumptions made by other classes.

Thanks to Brett Okken.
See: https://github.com/tukaani-project/xz-java/pull/18