SeekableXZInputStream.seekToBlock: Don't call public method getBlockPos
If a subclass overrides getBlockPos, it cannot affect seekToBlock now.
The call to getBlockPos was undocumented.

This is a change to the API/ABI. It's highly unlikely to affect
anything in practice because there's normally no reason to override
getBlockPos. If someone wanted to override getBlockPos, the call from
seekToBlock was likely an annoyance instead of a feature.