Remove downloading of "element-list" or "package-list"
"javadoc -linkoffline" needed a package list file to generate links
that point to the Java platform docs on docs.oracle.com. This kind
of build-time online dependency was always quite awkward but it
avoided committing such a file (or a custom shorter version of it)
into the repository.

Since OpenJDK 16, javadoc has been able to auto-link to the platform
docs without the need for -link or -linkoffline and its "package-list"
or "element-list" files. This means that it's possible generate links
to the platform docs when doing a fully offline build.

The downside is that links to platform docs won't be generated
anymore if building with OpenJDK 11 or 8.

Remove the "distclean" target as now "clean" already removes all
generated files.