build.xml: Force JVM locale to en_US when running javadoc
The docs itself are in English so it seems sensible if the generated
strings are always in English too (not depend on the locale of the
operating system). This helps with reproducible builds too.

javadoc has the option -locale and Ant exposes it as the "locale"
attribute. If one sets locale="en" or locale="en_US", it works
with OpenJDK 8. These commands produce identical HTML:

    LC_ALL=en_US.UTF-8 ant doc
    LC_ALL=ja_JP.UTF-8 ant doc

With OpenJDK 11, 17, and 22, it doesn't work. The only thing
that locale="en" or locale="en_US" changes is <html lang="ja">
to <html lang="en">. This isn't a new problem:

    https://bugs.openjdk.org/browse/JDK-8222793

Avoid javadoc's buggy -locale option. Instead, set the JVM default
locale for the javadoc process. This works with OpenJDK 8 too.