build.xml: Make MANIFEST.MF friendlier to reproducible builds
Override the values of Created-By and Ant-Version entries to make
reproducible builds easier. This can be disabled by setting
manifest_cleanup=false.

The JAR specification gives an impression that the Created-By
entry refers to the version and vendor of the Java implementation.
Ant's default output conforms to that. On the other hand, Maven
and its plugins use strings that refer to Maven or plugin names
and versions. Thus, in the real-world the Created-By entry can
refer to different things.