IA64.code: Silence a false positive from Coverity
bitRes is obviously at most 7 and thus (1 << bitRes) is at most 128,
so there is no risk of overflow. Coverity still doesn't like it because
the resulting int is widened to a long before the bitwise-and with instr.
Use (1L << bitRes) to avoid the widening and thus the warning.

Silences: Coverity CID 638721